cleanup(system): move pkg/cache to pkg/common/cache
