+++ pkg/common/cache/interface.go
